+++ ReleaseNotes.md
@@ -1,5 +1,9 @@
 Here's a list of the latest changes:
 
+**3.19.0**
+
+- [Enhancement - When organiser transfers a ticket to a new email, both original guest and new guest will receive a notification](https://trello.com/c/VtQklgSB/586-ticket-transfer-and-notification)
+
 **3.18.0**
 
 - [Change - Guests will be sent the ticket now](https://trello.com/c/r7biHc3s/584-each-guest-should-receive-the-ticket-along-with-an-invite)